warning and code clean up
