+++ src/components/CompassDetails/index.js
@@ -2,12 +2,10 @@ import React from "react";
 import { CompassViewerContainer, CompassNavigationBar } from "../../styles/CompassPage"
 import ProjectDetails from "./ProjectDetails";
 import ScribePermission from "./Permissions/ScribePermission";
-// import TeachersPermission from "./Permissions/TeachersPermission";
-import TeachersPermission from "./Permissions/NewTeachPermission";
-import MembersPermission from "./Permissions/MembersPermission";
-import ReadersPermission from "./Permissions/ReadersPermission";
+import TeachersPermission from "./Permissions/NewTeacherPermission";
+import MembersPermission from "./Permissions/NewMemberPermission";
+import ReadersPermission from "./Permissions/NewReaderPermission";
 import DisplayPermission from "./Permissions/DisplayPermission";
-// import DeletePermission from "./Permissions/DeletePermission";
 import { Box } from "grommet";
 import styled from "styled-components";
